Remove client.py
